Validate standard user transferring body user group access for detail browse views
